[Storage] Automatically Prune BlockStorage

+++ storage/block_storage.go
@@ -19,6 +19,7 @@ import (
 	"errors"
 	"fmt"
 	"log"
+	"strconv"
 
 	"github.com/coinbase/rosetta-sdk-go/syncer"
 	"github.com/coinbase/rosetta-sdk-go/types"
@@ -30,6 +31,10 @@ const (
 	// The head block is the block with the largest index that is
 	// not orphaned.
 	headBlockKey = "head-block"
+
+	// oldestBlockIndex is the last accessible block index. Anything
+	// prior to this block index has been pruned.
+	oldestBlockIndex = "oldest-block-index"
 
 	// blockNamespace is prepended to any stored block.
 	blockNamespace = "block"
@@ -49,6 +54,10 @@ type blockTransaction struct {
 
 func getHeadBlockKey() []byte {
 	return []byte(headBlockKey)
+}
+
+func getOldestBlockIndexKey() []byte {
+	return []byte(oldestBlockIndex)
 }
 
 func getBlockHashKey(hash string) (string, []byte) {
@@ -102,6 +111,173 @@ func NewBlockStorage(
 // This must be called prior to syncing!
 func (b *BlockStorage) Initialize(workers []BlockWorker) {
 	b.workers = workers
+}
+
+func (b *BlockStorage) setOldestBlockIndex(
+	ctx context.Context,
+	dbTx DatabaseTransaction,
+	update bool,
+	index int64,
+) error {
+	key := getOldestBlockIndexKey()
+	value := []byte(strconv.FormatInt(index, 10))
+	if update {
+		if err := dbTx.Set(ctx, key, value, true); err != nil {
+			return err
+		}
+
+		return nil
+	}
+
+	err := storeUniqueKey(ctx, dbTx, key, value, true)
+	if err == nil || errors.Is(err, ErrDuplicateKey) {
+		return nil
+	}
+
+	return err
+}
+
+// GetOldestBlockIndexTransactional returns the oldest block index
+// available in BlockStorage in a single database transaction.
+func (b *BlockStorage) GetOldestBlockIndexTransactional(
+	ctx context.Context,
+	dbTx DatabaseTransaction,
+) (int64, error) {
+	exists, rawIndex, err := dbTx.Get(ctx, getOldestBlockIndexKey())
+	if err != nil {
+		return -1, err
+	}
+
+	if !exists {
+		return -1, ErrOldestIndexMissing
+	}
+
+	index, err := strconv.ParseInt(string(rawIndex), 10, 64)
+	if err != nil {
+		return -1, err
+	}
+
+	return index, nil
+}
+
+// GetOldestBlockIndex returns the oldest block index
+// available in BlockStorage.
+func (b *BlockStorage) GetOldestBlockIndex(
+	ctx context.Context,
+) (int64, error) {
+	dbTx := b.db.NewDatabaseTransaction(ctx, false)
+	defer dbTx.Discard(ctx)
+
+	return b.GetOldestBlockIndexTransactional(ctx, dbTx)
+}
+
+// pruneBlock attempts to prune a single block in a database transaction.
+// If a block is pruned, we return its index.
+func (b *BlockStorage) pruneBlock(
+	ctx context.Context,
+	index int64,
+) (int64, error) {
+	// We create a separate transaction for each pruning attempt so that
+	// we don't hit the database tx size maximum. As a result, it is possible
+	// that we prune a collection of blocks, encounter an error, and cannot
+	// rollback the pruning operations.
+	dbTx := b.db.NewDatabaseTransaction(ctx, true)
+	defer dbTx.Discard(ctx)
+
+	oldestIndex, err := b.GetOldestBlockIndexTransactional(ctx, dbTx)
+	if err != nil {
+		return -1, fmt.Errorf("%w: %v", ErrOldestIndexRead, err)
+	}
+
+	if index < oldestIndex {
+		return -1, ErrNothingToPrune
+	}
+
+	head, err := b.GetHeadBlockIdentifierTransactional(ctx, dbTx)
+	if err != nil {
+		return -1, fmt.Errorf("%w: cannot get head block identifier", err)
+	}
+
+	if oldestIndex > head.Index-syncer.PastBlockSize*2 {
+		return -1, ErrPruningDepthInsufficient
+	}
+
+	blockResponse, err := b.getBlockResponse(
+		ctx,
+		&types.PartialBlockIdentifier{Index: &oldestIndex},
+		dbTx,
+	)
+	if err != nil && !errors.Is(err, ErrBlockNotFound) {
+		return -1, err
+	}
+
+	// If there is an omitted block, we will have a non-nil error. When
+	// a block is omitted, we should not attempt to remove it because
+	// it doesn't exist.
+	if err == nil {
+		blockIdentifier := blockResponse.Block.BlockIdentifier
+
+		for _, tx := range blockResponse.OtherTransactions {
+			if err := b.pruneTransaction(ctx, dbTx, blockIdentifier, tx); err != nil {
+				return -1, fmt.Errorf("%w: %v", ErrCannotPruneTransaction, err)
+			}
+		}
+
+		_, blockKey := getBlockHashKey(blockIdentifier.Hash)
+		if err := dbTx.Set(ctx, blockKey, []byte(""), true); err != nil {
+			return -1, err
+		}
+	}
+
+	// Update prune index
+	if err := b.setOldestBlockIndex(ctx, dbTx, true, oldestIndex+1); err != nil {
+		return -1, fmt.Errorf("%w: %v", ErrOldestIndexUpdateFailed, err)
+	}
+
+	// Commit tx
+	if err := dbTx.Commit(ctx); err != nil {
+		return -1, err
+	}
+
+	return oldestIndex, nil
+}
+
+// Prune removes block and transaction data
+// from all blocks with index <= index. Pruning
+// leaves all keys associated with pruned data
+// but overwrites their data to be empty. If pruning
+// is successful, we return the range of pruned
+// blocks.
+//
+// Prune is not invoked automatically because
+// some applications prefer not to prune any
+// block data.
+func (b *BlockStorage) Prune(
+	ctx context.Context,
+	index int64,
+) (int64, int64, error) {
+	firstPruned := int64(-1)
+	lastPruned := int64(-1)
+
+	for ctx.Err() == nil {
+		prunedBlock, err := b.pruneBlock(ctx, index)
+		if errors.Is(err, ErrNothingToPrune) {
+			return firstPruned, lastPruned, nil
+		}
+		if err != nil {
+			return -1, -1, fmt.Errorf("%w: %v", ErrPruningFailed, err)
+		}
+
+		if firstPruned == -1 {
+			firstPruned = prunedBlock
+		}
+
+		if lastPruned < prunedBlock {
+			lastPruned = prunedBlock
+		}
+	}
+
+	return -1, -1, ctx.Err()
 }
 
 // GetHeadBlockIdentifier returns the head block identifier,
@@ -183,7 +359,7 @@ func (b *BlockStorage) getBlockResponse(
 		// Get block by hash if provided
 		namespace, key = getBlockHashKey(*blockIdentifier.Hash)
 		exists, blockResponse, err = transaction.Get(ctx, key)
-	default:
+	case blockIdentifier.Index != nil:
 		// Get block by index if hash not provided
 		exists, key, err = transaction.Get(ctx, getBlockIndexKey(*blockIdentifier.Index))
 		namespace = blockIndexNamespace
@@ -197,7 +373,11 @@ func (b *BlockStorage) getBlockResponse(
 	}
 
 	if !exists {
-		return nil, fmt.Errorf("%w: %+v", ErrBlockNotFound, blockIdentifier)
+		return nil, fmt.Errorf("%w: %s", ErrBlockNotFound, types.PrintStruct(blockIdentifier))
+	}
+
+	if len(blockResponse) == 0 {
+		return nil, ErrCannotAccessPrunedData
 	}
 
 	var rosettaBlockResponse types.BlockResponse
@@ -223,18 +403,14 @@ func (b *BlockStorage) GetBlockLazy(
 	return b.getBlockResponse(ctx, blockIdentifier, transaction)
 }
 
-// GetBlock returns a block, if it exists. GetBlock
-// will fetch all transactions contained in a block
-// automatically. If you don't wish to do this for
-// performance reasons, use GetBlockLazy.
-func (b *BlockStorage) GetBlock(
+// GetBlockTransactional gets a block in the context of a database
+// transaction.
+func (b *BlockStorage) GetBlockTransactional(
 	ctx context.Context,
+	dbTx DatabaseTransaction,
 	blockIdentifier *types.PartialBlockIdentifier,
 ) (*types.Block, error) {
-	transaction := b.db.NewDatabaseTransaction(ctx, false)
-	defer transaction.Discard(ctx)
-
-	blockResponse, err := b.getBlockResponse(ctx, blockIdentifier, transaction)
+	blockResponse, err := b.getBlockResponse(ctx, blockIdentifier, dbTx)
 	if err != nil {
 		return nil, err
 	}
@@ -251,7 +427,7 @@ func (b *BlockStorage) GetBlock(
 			ctx,
 			block.BlockIdentifier,
 			transactionIdentifier,
-			transaction,
+			dbTx,
 		)
 		if err != nil {
 			return nil, fmt.Errorf(
@@ -267,6 +443,20 @@ func (b *BlockStorage) GetBlock(
 	block.Transactions = txs
 
 	return block, nil
+}
+
+// GetBlock returns a block, if it exists. GetBlock
+// will fetch all transactions contained in a block
+// automatically. If you don't wish to do this for
+// performance reasons, use GetBlockLazy.
+func (b *BlockStorage) GetBlock(
+	ctx context.Context,
+	blockIdentifier *types.PartialBlockIdentifier,
+) (*types.Block, error) {
+	transaction := b.db.NewDatabaseTransaction(ctx, false)
+	defer transaction.Discard(ctx)
+
+	return b.GetBlockTransactional(ctx, transaction, blockIdentifier)
 }
 
 func (b *BlockStorage) storeBlock(
@@ -297,6 +487,10 @@ func (b *BlockStorage) storeBlock(
 
 	if err := b.StoreHeadBlockIdentifier(ctx, transaction, blockIdentifier); err != nil {
 		return fmt.Errorf("%w: %v", ErrBlockIdentifierUpdateFailed, err)
+	}
+
+	if err := b.setOldestBlockIndex(ctx, transaction, false, blockIdentifier.Index); err != nil {
+		return fmt.Errorf("%w: %v", ErrOldestIndexUpdateFailed, err)
 	}
 
 	return nil
@@ -357,6 +551,19 @@ func (b *BlockStorage) deleteBlock(
 	block *types.Block,
 ) error {
 	blockIdentifier := block.BlockIdentifier
+
+	// We return an error if we attempt to remove the oldest index. If we did
+	// not error, it is possible that we could panic in the future as any
+	// further block removals would involve decoding pruned blocks.
+	oldestIndex, err := b.GetOldestBlockIndexTransactional(ctx, transaction)
+	if err != nil {
+		return fmt.Errorf("%w: %v", ErrOldestIndexRead, err)
+	}
+
+	if blockIdentifier.Index <= oldestIndex {
+		return ErrCannotRemoveOldest
+	}
+
 	_, key := getBlockHashKey(blockIdentifier.Hash)
 	if err := transaction.Delete(ctx, key); err != nil {
 		return fmt.Errorf("%w: %v", ErrBlockDeleteFailed, err)
@@ -381,13 +588,17 @@ func (b *BlockStorage) RemoveBlock(
 	ctx context.Context,
 	blockIdentifier *types.BlockIdentifier,
 ) error {
-	block, err := b.GetBlock(ctx, types.ConstructPartialBlockIdentifier(blockIdentifier))
+	transaction := b.db.NewDatabaseTransaction(ctx, true)
+	defer transaction.Discard(ctx)
+
+	block, err := b.GetBlockTransactional(
+		ctx,
+		transaction,
+		types.ConstructPartialBlockIdentifier(blockIdentifier),
+	)
 	if err != nil {
 		return err
 	}
-
-	transaction := b.db.NewDatabaseTransaction(ctx, true)
-	defer transaction.Discard(ctx)
 
 	// Remove all transaction hashes
 	for _, txn := range block.Transactions {
@@ -467,6 +678,24 @@ func (b *BlockStorage) SetNewStartIndex(
 		)
 	}
 
+	// Ensure we do not set a new start index less
+	// than the oldest block.
+	dbTx := b.db.NewDatabaseTransaction(ctx, false)
+	oldestIndex, err := b.GetOldestBlockIndexTransactional(ctx, dbTx)
+	dbTx.Discard(ctx)
+	if err != nil {
+		return fmt.Errorf("%w: %v", ErrOldestIndexRead, err)
+	}
+
+	if oldestIndex > startIndex {
+		return fmt.Errorf(
+			"%w: oldest block index is %d but start index is %d",
+			ErrCannotAccessPrunedData,
+			oldestIndex,
+			startIndex,
+		)
+	}
+
 	currBlock := head
 	for currBlock.Index >= startIndex {
 		log.Printf("Removing block %+v\n", currBlock)
@@ -514,23 +743,23 @@ func (b *BlockStorage) CreateBlockCache(ctx context.Context) []*types.BlockIdent
 	return cache
 }
 
-func storeUniqueKey(
+func (b *BlockStorage) updateTransaction(
 	ctx context.Context,
-	transaction DatabaseTransaction,
-	key []byte,
-	value []byte,
-	reclaimValue bool,
+	dbTx DatabaseTransaction,
+	hashKey []byte,
+	namespace string,
+	blocks map[string]*blockTransaction,
 ) error {
-	exists, _, err := transaction.Get(ctx, key)
+	encodedResult, err := b.db.Compressor().Encode(namespace, blocks)
 	if err != nil {
+		return fmt.Errorf("%w: %v", ErrTransactionDataEncodeFailed, err)
+	}
+
+	if err := dbTx.Set(ctx, hashKey, encodedResult, true); err != nil {
 		return err
 	}
 
-	if exists {
-		return fmt.Errorf("%w: duplicate key %s found", ErrDuplicateKey, string(key))
-	}
-
-	return transaction.Set(ctx, key, value, reclaimValue)
+	return nil
 }
 
 func (b *BlockStorage) storeTransaction(
@@ -568,16 +797,34 @@ func (b *BlockStorage) storeTransaction(
 		BlockIndex:  blockIdentifier.Index,
 	}
 
-	encodedResult, err := b.db.Compressor().Encode(namespace, blocks)
-	if err != nil {
-		return fmt.Errorf("%w: %v", ErrTransactionDataEncodeFailed, err)
-	}
+	return b.updateTransaction(ctx, transaction, hashKey, namespace, blocks)
+}
 
-	if err := transaction.Set(ctx, hashKey, encodedResult, true); err != nil {
+func (b *BlockStorage) pruneTransaction(
+	ctx context.Context,
+	transaction DatabaseTransaction,
+	blockIdentifier *types.BlockIdentifier,
+	txIdentifier *types.TransactionIdentifier,
+) error {
+	namespace, hashKey := getTransactionHashKey(txIdentifier)
+	exists, val, err := transaction.Get(ctx, hashKey)
+	if err != nil {
 		return err
 	}
+	if !exists {
+		return ErrTransactionNotFound
+	}
 
-	return nil
+	var blocks map[string]*blockTransaction
+	if err := b.db.Compressor().Decode(namespace, val, &blocks, true); err != nil {
+		return fmt.Errorf("%w: could not decode transaction hash contents", err)
+	}
+
+	blocks[blockIdentifier.Hash] = &blockTransaction{
+		BlockIndex: blockIdentifier.Index,
+	}
+
+	return b.updateTransaction(ctx, transaction, hashKey, namespace, blocks)
 }
 
 func (b *BlockStorage) removeTransaction(
@@ -611,16 +858,7 @@ func (b *BlockStorage) removeTransaction(
 		return transaction.Delete(ctx, hashKey)
 	}
 
-	encodedResult, err := b.db.Compressor().Encode(namespace, blocks)
-	if err != nil {
-		return fmt.Errorf("%w: %v", ErrTransactionDataEncodeFailed, err)
-	}
-
-	if err := transaction.Set(ctx, hashKey, encodedResult, true); err != nil {
-		return err
-	}
-
-	return nil
+	return b.updateTransaction(ctx, transaction, hashKey, namespace, blocks)
 }
 
 // FindTransaction returns the most recent *types.BlockIdentifier containing the
@@ -655,6 +893,11 @@ func (b *BlockStorage) FindTransaction(
 		}
 	}
 
+	// If the transaction has been pruned, it will be nil.
+	if newestTransaction == nil {
+		return nil, nil, ErrCannotAccessPrunedData
+	}
+
 	return newestBlock, newestTransaction, nil
 }
 
@@ -664,6 +907,15 @@ func (b *BlockStorage) findBlockTransaction(
 	transactionIdentifier *types.TransactionIdentifier,
 	txn DatabaseTransaction,
 ) (*types.Transaction, error) {
+	oldestIndex, err := b.GetOldestBlockIndexTransactional(ctx, txn)
+	if err != nil {
+		return nil, fmt.Errorf("%w: %v", ErrOldestIndexRead, err)
+	}
+
+	if blockIdentifier.Index < oldestIndex {
+		return nil, ErrCannotAccessPrunedData
+	}
+
 	namespace, key := getTransactionHashKey(transactionIdentifier)
 	txExists, tx, err := txn.Get(ctx, key)
 	if err != nil {
